fix: remove mock implementation and interface declaration in a separate file and put interface and implementation in the xcvbnm.dart file
