+++ test/scoring_test.dart
@@ -4,7 +4,7 @@ import 'package:test/test.dart';
 import 'dart:math' as math;
 import 'package:xcvbnm/src/scoring.dart';
 import 'package:xcvbnm/src/matching.dart' as matching;
-import 'package:xcvbnm/src/xcvbnm_common.dart' as xcvbnm;
+import 'package:xcvbnm/xcvbnm.dart' as xcvbnm;
 import 'dart:core' hide Match;
 
 main() {
